Move seriesForPostingsCacheEntry and util functions to series_refs.go
